Update TurretCannon Target List
Instead of creating a new target list on every timer refresh, now there is an array that stores the results in the class.

This means that if I need to access the targets, I can use the list which will update automatically every 1s instead of generating a new sorted list.

Updated method names to reduce ambiguity.