Added Currency System + Updated SpaceStation UI to handle it
Added a currency system which uses a resource.

The space station UI has a copy of the player resource which it gives a reference to each turret control. Each turret control can then spend/add money to the resource.

To ensure that the UI is not ready before the space station, a new signal has been added to the space station which is emitted when it is ready. The UI then gets the path of the space station, and stores a reference to the turret manager.