Engine Thrust Particles + Turret Muzzle Particles + TextureProgress Updates & FIxes
Added particles to the engine thruster of the player., as well as to the muzzle of the cannons when they shoot.

Fixed the health, fuel and base health bars so that progress on them now appears correctly. Moved fuel bar to be near to the player, to match the health bar.