Added hit shader + Split beam spritesheet into two seperate
Added a hit shader which can be used instead of manually drawing the hit effect on a sprite.

Split the beam spritesheet into the two separate parts, this should make it easier to animate as the keyframes for turret & base will be the same, and also ensures no frames for the other part appear as part of the other.