Added toggle for ship brake + Brake HUD + refactored player script
Added a toggle for the ship brake so that it can be turned on or off.

Added an indicator for the HUD when the brake is turned on.

Refactored names of variables in the player script. Private variables are now prefixed with an underscore to show they should not be accessed outside of the class.